cobalt/synch: test information bits as returned by xnsynch_sleep_on() only
For consistency with xnsynch_acquire(), callers of xnsynch_sleep_on()
MUST retrieve the information bits from the return value, instead of
peeking into the current thread's TCB.

+++ kernel/cobalt/synch.c
@@ -189,7 +189,7 @@ int xnsynch_sleep_on(struct xnsynch *synch, xnticks_t timeout,
 
 	xnlock_put_irqrestore(&nklock, s);
 
-	return (int)xnthread_test_info(thread, XNRMID|XNTIMEO|XNBREAK);
+	return xnthread_test_info(thread, XNRMID|XNTIMEO|XNBREAK);
 }
 EXPORT_SYMBOL_GPL(xnsynch_sleep_on);
